Updated the flight component with view. Store now fetches the data from repository[should be changed - temporarily]

+++ features/src/main/java/com/kinley/features/flights/presentation/FlightUiModel.kt
@@ -6,5 +6,5 @@ package com.kinley.features.flights.presentation
 data class FlightUiModel(
     val flightName: String,
     val airline: String,
-    val flightCost: Double
+    val flightCost: Int
 )
